Tweak blocks_for_region to omit hidden settings/nav blocks
Settings and navigation blocks would only be hidden by this function if they were not set to hidden; otherwise, they would appear again when editing mode was turned on.  This change may also help boost performance of this function slightly, since it now checks the setting before checking whether it's the block that the setting applies to (hooray for shortcut operators).