Edit buttons: Update captions on show/hide and group mode when clicked
The click event fires before the icon has changed, so setTimeout is required to make it fire after the change.

Event with a 1ms timeout requested, the AJAX request seems to block it - so in some browsers, the caption update lags behind the icon change noticeably.

Also changes the icon URL matching to identify these icons, as "&svg=0" is appended to the URLs in IE(8).