modchooser tiles: Hide help text when in phone mode
At the smallest breakpoint, the help text gets mangled by an attempt to
display it below the option, which doesn't work in the tile layout.  Hide
the help entirely in this view; if the user really needs it on mobile,
they can rotate to landscape mode to see the side-by-side help.