Shift completion checkboxes to outside the list of CMs
This tidies up the completion checkboxes, putting them in a consistent
location so that they can easily be seen by both teachers and students.
Previously, it was on the opposite side for labels - this also tidies that
up, putting it in the same place for all types of activity/resource.
Includes RTL support - completion checkboxes go on the right-hand side in
RTL, so they're still consistently beside the module type icons.